requirements.txt
web3
python-dotenv
rich
requests
